backend/requirements.txt
Flask==3.1.0
flask-cors==5.0.1
flask-socketio==5.4.1
flask-sqlalchemy==3.1.1
flask-migrate==4.1.0
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
redis==5.2.1
requests==2.32.3
praw==7.8.1
apscheduler==3.11.0
python-dotenv==1.0.1
gunicorn==23.0.0
eventlet==0.37.0
httpx==0.28.1
PyJWT==2.10.1
cryptography==44.0.0
